Organize project into package structure
