+++ README.md
@@ -1,2 +1,10 @@
 # stell-structure-optimizer
 Analyzer and Optimizer for Steel Structures Using Commercial Steel Profiles
+
+## Installation
+
+Install the package using `pip` from the repository root:
+
+```bash
+pip install .
+```
