+++ requirements.txt
@@ -1,0 +1,6 @@
+numpy
+pandas
+scipy
+matplotlib
+plotly
+streamlit
