+++ src/stell_structure_optimizer/__init__.py
@@ -1,0 +1,1 @@
+"""Steel structure optimization tools"""
