+++ src/stell_structure_optimizer/analysis/__init__.py
@@ -1,0 +1,1 @@
+
