+++ src/stell_structure_optimizer/analysis/structural_model.py
