+++ src/stell_structure_optimizer/data/profile_manager.py
